Security: run scripts as non-root user
These scripts don't need to be run as root; run them as a new system
user called pe_metric_curl_cron_jobs instead.

This enforces ownership for a number of files to ensure they're not
writable by the scripts. It also moves the scripts directory out of the
output directory to ensure that directory tree below scripts is not
writable by the pe_metric_curl_cron_jobs user.